+++ Js2IL.Tests/BinaryOperator/ExecutionTests.cs
@@ -60,6 +60,9 @@ namespace Js2IL.Tests.BinaryOperator
         public Task BinaryOperator_LeftShiftNumberNumber() { var testName = nameof(BinaryOperator_LeftShiftNumberNumber); return ExecutionTest(testName); }
 
         [Fact]
+        public Task BinaryOperator_LeftShiftBit31() { var testName = nameof(BinaryOperator_LeftShiftBit31); return ExecutionTest(testName); }
+
+        [Fact]
         public Task BinaryOperator_LessThan() { var testName = nameof(BinaryOperator_LessThan); return ExecutionTest(testName); }
 
         [Fact]